Add lunr search
Implemented an initial search page based on the lunr plugin.

TODO: change the lunr resolver for the markdown body when it's
indexing in gatsby-config.js to strip the markdown. This will
make the search only search the text and will help allow the
search page to truncate the body result matches more cleanly.